chore: Typescript-react 연습
- 이걸 토대로 개발해나갈 예정..
