+++ fe/src/components/Main.tsx
@@ -1,0 +1,27 @@
+import React, { useState } from 'react';
+import Hero from './mainComponents/Hero';
+import NearbyLocation from './mainComponents/NearbyLocation';
+import HouseType from './mainComponents/HouseType';
+import BeingHost from './mainComponents/BeingHost';
+
+interface User {
+    email: string;
+    name: string;
+}
+
+const Main = (): React.ReactElement => {
+    const [user, setUser] = useState<User>({
+        email: '',
+        name: '',
+    });
+    return (
+        <>
+            <Hero user={user} setUser={setUser} />
+            <NearbyLocation />
+            <HouseType />
+            <BeingHost />
+        </>
+    );
+};
+
+export default Main;
