--- fe/src/components/mainComponents/BeingHost.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const BeingHost = (props) => {
-  return <div>BeingHost</div>;
-};
-
-export default BeingHost;
