+++ fe/src/components/mainComponents/BeingHost.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const BeingHost = (): React.ReactElement => {
+    return <div>BeingHost</div>;
+};
+
+export default BeingHost;
