--- fe/src/components/mainComponents/HouseType.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const HouseType = (props) => {
-    return (
-        <div>HouseType</div>
-    )
-};
-
-export default HouseType;
