+++ fe/src/components/mainComponents/HouseType.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const HouseType = (): React.ReactElement => {
+    return <div>HouseType</div>;
+};
+
+export default HouseType;
