--- fe/src/components/mainComponents/NearbyLocation.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const NearbyLocation = (props) => {
-  return <div>NearbyLocation</div>;
-};
-
-export default NearbyLocation;
