+++ fe/src/components/mainComponents/NearbyLocation.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const NearbyLocation = (): React.ReactElement => {
+    return <div>NearbyLocation</div>;
+};
+
+export default NearbyLocation;
