+++ fe/text.bat
@@ -1,0 +1,1 @@
+ren *.jsx *.tsx
